refactor(galgame): move vision training into plugin
