+++ plugin/plugins/galgame_plugin/training/__init__.py
