+++ plugin/plugins/galgame_plugin/training/classify/config.yaml
