+++ plugin/plugins/galgame_plugin/training/classify/eval.py
@@ -1,11 +1,11 @@
-from __future__ import annotations
+﻿from __future__ import annotations
 
 import json
 from pathlib import Path
 
 import numpy as np
 
-from training.shared.metrics import macro_f1, top1_accuracy
+from plugin.plugins.galgame_plugin.training.shared.metrics import macro_f1, top1_accuracy
 
 
 def write_eval_report(logits: np.ndarray, labels: np.ndarray, output_path: str | Path) -> dict[str, float]:
