+++ plugin/plugins/galgame_plugin/training/data/__init__.py
