+++ plugin/plugins/galgame_plugin/training/data/raw/.gitkeep
