+++ plugin/plugins/galgame_plugin/training/data/test_fixtures/.gitkeep
