+++ plugin/plugins/galgame_plugin/training/pyproject.toml
