+++ plugin/plugins/galgame_plugin/training/screenshots/.gitkeep
