+++ plugin/plugins/galgame_plugin/training/shared/__init__.py
